added search route, refactored routes, fixed some issues
