+++ controller/users.js
@@ -58,6 +58,8 @@ export const loginUser = async (req, res) => {
   }
 }
 
+
+
 export const getUserPreferences = async (req, res) => {
   const { email } = req.user;
   try {
@@ -101,199 +103,3 @@ export const addUserPreferences = async (req, res) => {
   }
 }
 
-export const getFavoriteNews = async (req, res) => {
-  const { email } = req.user;
-  try {
-    const user = await User
-      .findOne({ email })
-      .exec();
-    if (!user) {
-      return res.status(400).json({ message: 'User not found' });
-    }
-    if (user.favorites.length === 0) {
-      return res.status(200).json({ message: 'No favorites found' });
-    }
-    return res.status(200).json({ preferences: user.favorites });
-
-  } catch (error) {
-    console.error(error);
-    return res.status(500).json({ message: 'Internal server error' });
-
-  }
-}
-
-export const addFavoriteNews = async (req, res) => {
-  const { email } = req.user;
-  const { articleId } = req.body;
-  if (!articleId) {
-    return res.status(400).json({ message: 'Please provide articleId' });
-  }
-  try {
-    const user = await User
-      .findOne({ email })
-      .exec();
-    if (!user) {
-      return res.status(400).json({ message: 'User not found' });
-    }
-
-    user.favorites.push(articleId);
-    await user.save();
-    return res.status(200).json({ message: 'Article added to favorites successfully' });
-  } catch (error) {
-    console.error(error);
-    return res.status(500).json({ message: 'Internal server error' });
-  }
-}
-export const removeFavoriteNews = async (req, res) => {
-  const { email } = req.user;
-  const { articleId } = req.body;
-  if (!articleId) {
-    return res.status(400).json({ message: 'Please provide articleId' });
-  }
-  try {
-    const user = await User
-      .findOne({ email })
-      .exec();
-    if (!user) {
-      return res.status(400).json({ message: 'User not found' });
-    }
-
-    user.favorites = user.favorites.filter(id => id !== articleId);
-    await user.save();
-    return res.status(200).json({ message: 'Article removed from favorites successfully' });
-  } catch (error) {
-    console.error(error);
-    return res.status(500).json({ message: 'Internal server error' });
-  }
-}
-
-export const getReadNews = async (req, res) => {
-  const { email } = req.user;
-  try {
-    const user = await User
-      .findOne({ email })
-      .exec();
-    if (!user) {
-      return res.status(400).json({ message: 'User not found' });
-    }
-    if (user.favorites.length === 0) {
-      return res.status(200).json({ message: 'No read found' });
-    }
-    return res.status(200).json({ read: user.read });
-  } catch (error) {
-    console.error(error);
-    return res.status(500).json({ message: 'Internal server error' });
-
-  }
-}
-
-export const addReadNews = async (req, res) => {
-  const { email } = req.user;
-  const { articleId } = req.body;
-  if (!articleId) {
-    return res.status(400).json({ message: 'Please provide articleId' });
-  }
-  try {
-    const user = await User
-      .findOne({ email })
-      .exec();
-    if (!user) {
-      return res.status(400).json({ message: 'User not found' });
-    }
-
-    user.read.push(articleId);
-    await user.save();
-    return res.status(200).json({ message: 'Article added to read successfully' });
-  } catch (error) {
-    console.error(error);
-    return res.status(500).json({ message: 'Internal server error' });
-  }
-}
-
-export const removeReadNews = async (req, res) => {
-  const { email } = req.user;
-  const { articleId } = req.body;
-  if (!articleId) {
-    return res.status(400).json({ message: 'Please provide articleId' });
-  }
-  try {
-    const user = await User
-      .findOne({ email })
-      .exec();
-    if (!user) {
-      return res.status(400).json({ message: 'User not found' });
-    }
-
-    user.read = user.read.filter(id => id !== articleId);
-    await user.save();
-    return res.status(200).json({ message: 'Article removed from read successfully' });
-  } catch (error) {
-    console.error(error);
-    return res.status(500).json({ message: 'Internal server error' });
-  }
-}
-
-export const getSavedNews = async (req, res) => {
-  const { email } = req.user;
-  try {
-    const user = await User
-      .findOne({ email })
-      .exec();
-    if (!user) {
-      return res.status(400).json({ message: 'User not found' });
-    }
-    if (user.favorites.length === 0) {
-      return res.status(200).json({ message: 'No saved found' });
-    }
-    return res.status(200).json({ read: user.saved });
-  } catch (error) {
-    console.error(error);
-    return res.status(500).json({ message: 'Internal server error' });
-  }
-}
-
-export const addSavedNews = async (req, res) => {
-  const { email } = req.user;
-  const { articleId } = req.body;
-  if (!articleId) {
-    return res.status(400).json({ message: 'Please provide articleId' });
-  }
-  try {
-    const user = await User
-      .findOne({ email })
-      .exec();
-    if (!user) {
-      return res.status(400).json({ message: 'User not found' });
-    }
-
-    user.saved.push(articleId);
-    await user.save();
-    return res.status(200).json({ message: 'Article added to saved successfully' });
-  } catch (error) {
-    console.error(error);
-    return res.status(500).json({ message: 'Internal server error' });
-  }
-}
-
-export const removeSavedNews = async (req, res) => {
-  const { email } = req.user;
-  const { articleId } = req.body;
-  if (!articleId) {
-    return res.status(400).json({ message: 'Please provide articleId' });
-  }
-  try {
-    const user = await User
-      .findOne({ email })
-      .exec();
-    if (!user) {
-      return res.status(400).json({ message: 'User not found' });
-    }
-
-    user.saved = user.saved.filter(id => id !== articleId);
-    await user.save();
-    return res.status(200).json({ message: 'Article removed from saved successfully' });
-  } catch (error) {
-    console.error(error);
-    return res.status(500).json({ message: 'Internal server error' });
-  }
-}
